+++ lib/src/bones_api_entity.dart
@@ -30,6 +30,7 @@ import 'bones_api_utils.dart';
 import 'bones_api_utils_collections.dart';
 import 'bones_api_utils_instance_tracker.dart';
 import 'bones_api_utils_json.dart';
+import 'bones_api_utils_weaklist.dart';
 
 final _log = logging.Logger('Entity');
 
@@ -480,6 +481,14 @@ abstract class EntityHandler<O> with FieldsFromMap, EntityRulesResolver {
                   isValidEntityType(fieldType.arguments0!.type)),
           o);
 
+  Map<String, TypeInfo>? _fieldsWithEntityReference;
+
+  Map<String, TypeInfo> fieldsWithEntityReference([O? o]) =>
+      _fieldsWithEntityReference ??= fieldsWithTypeEntityOrReference(o)
+          .entries
+          .where((e) => e.value.isEntityReferenceType)
+          .toMapFromEntries();
+
   Map<String, TypeInfo>? _fieldsWithTypeListEntity;
 
   Map<String, TypeInfo> fieldsWithTypeListEntity([O? o]) =>
@@ -491,6 +500,14 @@ abstract class EntityHandler<O> with FieldsFromMap, EntityRulesResolver {
   Map<String, TypeInfo> fieldsWithTypeListEntityOrReference([O? o]) =>
       _fieldsWithTypeListEntityReference ??= fieldsWithType(
           (_, fieldType) => fieldType.isValidListEntityOrReferenceType, o);
+
+  Map<String, TypeInfo>? _fieldsWithEntityReferenceList;
+
+  Map<String, TypeInfo> fieldsWithEntityReferenceList([O? o]) =>
+      _fieldsWithEntityReferenceList ??= fieldsWithTypeListEntityOrReference(o)
+          .entries
+          .where((e) => e.value.isEntityReferenceListType)
+          .toMapFromEntries();
 
   Map<String, TypeInfo> fieldsWithType(
       bool Function(String fieldName, TypeInfo fieldType) typeFilter,
@@ -2473,6 +2490,8 @@ abstract class EntityAccessor<O extends Object> {
 abstract class EntitySource<O extends Object> extends EntityAccessor<O> {
   EntitySource(String name) : super(name);
 
+  bool hasReferencedEntities();
+
   FutureOr<bool> existsID(dynamic id, {Transaction? transaction});
 
   FutureOr<O?> selectByID(dynamic id, {Transaction? transaction}) {
@@ -4051,19 +4070,35 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
   /// See [execute] and [executeBlock].
   static Transaction? get executingTransaction => _executingTransaction.get();
 
+  static final WeakList<Transaction> _openInstances = WeakList<Transaction>();
+
+  /// Returns the `Transaction` that are open.
+  static List<Transaction> get openInstances => _openInstances.toList();
+
   final List<TransactionOperation> _operations = <TransactionOperation>[];
 
   List<TransactionOperation> get operations =>
       UnmodifiableListView(_operations);
 
-  Object? get executor => _operations.firstOrNull?.executor;
+  /// Returns the 1st [operations].
+  TransactionOperation? get mainOperation => _operations.firstOrNull;
+
+  /// Returns `canPropagate` of the [mainOperation].
+  /// See [TransactionOperation.canPropagate].
+  bool get canPropagate => mainOperation?.canPropagate ?? false;
+
+  /// Returns the executor of this transaction.
+  Object? get executor => mainOperation?.executor;
 
   final List<TransactionOperation> _executedOperations =
       <TransactionOperation>[];
 
+  /// Returns the list of [operations] that were fully executed.
   List<TransactionOperation> get executedOperations =>
       UnmodifiableListView(_executedOperations);
 
+  /// Returns true if this transaction will automatically commit once
+  /// all operations have been executed.
   final bool autoCommit;
 
   late final Completer _transactionCompleter;
@@ -4185,15 +4220,19 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
 
   bool get isCommitting => _commitCalled && (!_committed && !_aborted);
 
+  /// Returns the [operations] length.
   int get length => _operations.length;
 
+  /// Returns `true` if this transaction doesn't have [operations].
   bool get isEmpty => _operations.isEmpty;
 
+  /// Alias to ![isEmpty].
   bool get isNotEmpty => _operations.isNotEmpty;
 
   int get notExecutedOperationsSize =>
       _operations.length - _executedOperations.length;
 
+  /// Returns `true` if [op] is being executed.
   bool isLastExecutingOperation(TransactionOperation op) {
     if (isEmpty) return false;
 
@@ -4215,10 +4254,12 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
 
   bool _opening = false;
 
+  /// Returns `true` if this transaction is in the middle of the `open` process.
   bool get isOpening => _opening;
 
   FutureOr<void> Function()? _transactionCloser;
 
+  /// Opens this transaction. This is called by the [DBAdapter] implementation.
   FutureOr<Object?> open(
       FutureOr<Object> Function() opener, FutureOr<void> Function()? closer) {
     if (_opening) {
@@ -4226,6 +4267,8 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
     } else if (_open) {
       throw StateError("Transaction already open:\n$this");
     }
+
+    _initTime ??= DateTime.now();
 
     _opening = true;
     _transactionCloser = closer;
@@ -4238,8 +4281,26 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
     });
   }
 
+  DateTime? _initTime;
+
+  /// The time of transaction initialization.
+  DateTime? get initTime => _initTime;
+
+  DateTime? _endTime;
+
+  /// The time of transaction completion.
+  DateTime? get endTime => _endTime;
+
+  /// The duration of the transaction execution.
+  Duration? get duration {
+    var init = _initTime;
+    var end = _endTime;
+    return init != null && end != null ? end.difference(init) : null;
+  }
+
   bool _open = false;
 
+  /// Returns `true` if this transaction was open.
   bool get isOpen => _open;
 
   Object? _context;
@@ -4251,6 +4312,8 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
     _open = true;
     _opening = false;
     _openCompleter.complete(true);
+
+    _openInstances.add(this);
   }
 
   FutureOr<R> onOpen<R>(FutureOr<R> Function() f) {
@@ -4267,6 +4330,7 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
 
   Object? _result;
 
+  /// Returns the final result of this transaction.
   Object? get result => _result;
 
   void addOperation(TransactionOperation op) {
@@ -4285,6 +4349,8 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
     if (_commitCalled) {
       throw StateError("Transaction is committing:\n$this");
     }
+
+    _initTime ??= DateTime.now();
 
     _operations.add(op);
     op.id = _operations.length;
@@ -4332,6 +4398,8 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
       throw StateError("Transaction already committed:\n$this");
     }
 
+    op._endTime = DateTime.now();
+
     _executedOperations.add(op);
 
     if (!_aborted) {
@@ -4352,6 +4420,7 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
     return null;
   }
 
+  /// Waits for all [operation]s to execute.
   FutureOr<bool> waitAllExecuted() {
     if (_executedOperations.length == _operations.length) {
       return true;
@@ -4378,6 +4447,7 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
 
   bool _commitCalled = false;
 
+  /// Commits this transaction, returning the final result.
   FutureOr<R?> commit<R>() {
     if (_aborted) {
       if (_commitCalled) {
@@ -4465,13 +4535,44 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
       }
     }
 
+    _endTime ??= DateTime.now();
+
     _committed = true;
 
-    _logTransaction.info(
-        '[transaction:$id] Committed> ops: ${_operations.length} ; root: ${_operations.firstOrNull} > result: $result');
+    _close();
 
     _resultCompleter.complete(result);
+
+    if (_logTransaction.isLoggable(logging.Level.INFO)) {
+      var duration = this.duration ?? Duration(milliseconds: -1);
+
+      _logTransaction.info(
+          '[transaction:$id] Committed> time: ${duration
+              .inMilliseconds} ms ;  ops: ${_operations
+              .length} ; root: ${_operations.firstOrNull} > result: $result');
+
+      if (duration.inMilliseconds > 500) {
+        _logTransaction.warning(
+            'SLOW TRANSACTION (${duration.inMilliseconds} ms):\n${toString(
+                withExecutedOperations: false)}');
+      } else if (_operations.length > 15) {
+        _logTransaction.warning(
+            'LONG TRANSACTION (${_operations.length}):\n${toString(
+                withExecutedOperations: false)}');
+      }
+    }
+
     return result as R?;
+  }
+
+  void _close({Zone? zone}) {
+    zone ??= Zone.current;
+
+    if (identical(_executingTransaction.get(zone), this)) {
+      _executingTransaction.disposeContextZone(zone);
+    }
+
+    _openInstances.remove(this);
   }
 
   bool _aborted = false;
@@ -4484,7 +4585,7 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
   /// Returns the abort error ([TransactionAbortedError]).
   TransactionAbortedError? get abortedError => _abortedError;
 
-  /// Abort this transaction.
+  /// Aborts this transaction.
   FutureOr<TransactionAbortedError?> abort(
       {String? reason,
       Object? payload,
@@ -4541,6 +4642,8 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
 
     _transactionCompleter.completeError(
         abortError.abortError ?? abortError, abortError.abortStackTrace);
+
+    _close();
 
     if (transactionResult != null) {
       if (transactionResult is Future) {
@@ -4618,11 +4721,7 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
       _onExecutionError(error, stackTrace, null, null, rethrowError: false);
     }
 
-    return asyncTry(commit, onFinally: () {
-      if (identical(_executingTransaction.get(zone), this)) {
-        _executingTransaction.disposeContextZone(zone);
-      }
-    });
+    return asyncTry(commit, onFinally: () => _close(zone: zone));
   }
 
   final List<FutureOr> _executionsFutures = <FutureOr>[];
@@ -4732,7 +4831,12 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
   }
 
   @override
-  String toString({bool compact = false}) {
+  String toString(
+      {bool compact = false,
+      bool withOperations = true,
+      bool withExecutedOperations = true}) {
+    final duration = this.duration;
+
     return [
       'Transaction[#$id]{\n',
       '  open: $isOpen\n',
@@ -4742,16 +4846,21 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
       '  abortedError: $abortedError\n',
       '  cachedEntities: $cachedEntitiesLength\n',
       '  external: $_external\n',
-      if (compact) '  operations: ${operations.length}\n',
-      if (!compact) '  operations: [\n',
-      if (!compact && _operations.isNotEmpty)
-        '    ${_operations.join(',\n    ')}',
-      if (!compact) '\n  ],\n',
-      if (compact) '  executedOperations: ${executedOperations.length}\n',
-      if (!compact) '  executedOperations: [\n',
-      if (!compact && _executedOperations.isNotEmpty)
-        '    ${_executedOperations.join(',\n    ')}',
-      if (!compact) '\n  ]\n',
+      if (duration != null) '  duration: ${duration.inMilliseconds} ms\n',
+      if (withOperations) ...[
+        if (compact) '  operations: ${operations.length}\n',
+        if (!compact) '  operations: [\n',
+        if (!compact && _operations.isNotEmpty)
+          '    ${_operations.join(',\n    ')}',
+        if (!compact) '\n  ],\n',
+      ],
+      if (withExecutedOperations) ...[
+        if (compact) '  executedOperations: ${executedOperations.length}\n',
+        if (!compact) '  executedOperations: [\n',
+        if (!compact && _executedOperations.isNotEmpty)
+          '    ${_executedOperations.join(',\n    ')}',
+        if (!compact) '\n  ]\n',
+      ],
       if (_commitCalled) '  result: $_result\n',
       '}'
     ].join();
@@ -4761,6 +4870,11 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
 abstract class TransactionOperation {
   final TransactionOperationType type;
   final String repositoryName;
+
+  /// If `true` indicates that this operation can generated multiple operations,
+  /// usually sub-entities requests/resolution.
+  final bool canPropagate;
+
   Object? command;
 
   int? id;
@@ -4773,8 +4887,10 @@ abstract class TransactionOperation {
   late final bool externalTransaction;
   late final bool transactionRoot;
 
-  TransactionOperation(
-      this.type, this.repositoryName, this.executor, Transaction? transaction) {
+  final DateTime initTime = DateTime.now();
+
+  TransactionOperation(this.type, this.repositoryName, this.canPropagate,
+      this.executor, Transaction? transaction) {
     var resolvedTransaction =
         resolveTransaction(transaction: transaction, operation: this);
     this.transaction = resolvedTransaction;
@@ -4828,6 +4944,17 @@ abstract class TransactionOperation {
     return transaction;
   }
 
+  DateTime? _endTime;
+
+  DateTime? get endTime => _endTime;
+
+  Duration? get duration => _endTime?.difference(initTime);
+
+  String get durationMsInfo {
+    var duration = this.duration;
+    return duration == null ? '' : '(${duration.inMilliseconds} ms)';
+  }
+
   int get transactionId => transaction.id;
 
   TransactionExecution? execution;
@@ -4835,6 +4962,9 @@ abstract class TransactionOperation {
   String get _commandToString => command == null ? '' : ', command: $command';
 
   FutureOr<R> finish<R>(R result) => transaction.finishOperation(this, result);
+
+  @override
+  String toString();
 }
 
 class TransactionOperationSubTransaction<O> extends TransactionOperation {
@@ -4845,12 +4975,12 @@ class TransactionOperationSubTransaction<O> extends TransactionOperation {
     Object executor,
     Transaction parentTransaction,
     this.subTransaction,
-  ) : super(TransactionOperationType.subTransaction, repositoryName, executor,
-            parentTransaction);
+  ) : super(TransactionOperationType.subTransaction, repositoryName, true,
+            executor, parentTransaction);
 
   @override
   String toString() {
-    return 'TransactionOperationSubTransaction[#$id@#${transaction.id}:subTransaction@$repositoryName]->${subTransaction.toString(compact: true)}';
+    return 'TransactionOperationSubTransaction[#$id@#${transaction.id}:subTransaction@$repositoryName]->${subTransaction.toString(compact: true)}$durationMsInfo';
   }
 }
 
@@ -4858,14 +4988,14 @@ class TransactionOperationSelect<O> extends TransactionOperation {
   final EntityMatcher matcher;
 
   TransactionOperationSelect(
-      String repositoryName, Object executor, this.matcher,
+      String repositoryName, bool canPropagate, Object executor, this.matcher,
       [Transaction? transaction])
-      : super(TransactionOperationType.select, repositoryName, executor,
-            transaction);
+      : super(TransactionOperationType.select, repositoryName, canPropagate,
+            executor, transaction);
 
   @override
   String toString() {
-    return 'TransactionOperation[#$id@#${transaction.id}:select@$repositoryName]{matcher: $matcher$_commandToString}';
+    return 'TransactionOperation[#$id@#${transaction.id}:select@$repositoryName]{matcher: $matcher$_commandToString}$durationMsInfo';
   }
 }
 
@@ -4874,26 +5004,27 @@ class TransactionOperationCount<O> extends TransactionOperation {
 
   TransactionOperationCount(String repositoryName, Object executor,
       [this.matcher, Transaction? transaction])
-      : super(TransactionOperationType.count, repositoryName, executor,
+      : super(TransactionOperationType.count, repositoryName, false, executor,
             transaction);
 
   @override
   String toString() {
-    return 'TransactionOperation[#$id@#${transaction.id}:count@$repositoryName]{matcher: $matcher$_commandToString}';
+    return 'TransactionOperation[#$id@#${transaction.id}:count@$repositoryName]{matcher: $matcher$_commandToString}$durationMsInfo';
   }
 }
 
 class TransactionOperationStore<O> extends TransactionOperation {
   final O entity;
 
-  TransactionOperationStore(String repositoryName, Object executor, this.entity,
+  TransactionOperationStore(
+      String repositoryName, bool canPropagate, Object executor, this.entity,
       [Transaction? transaction])
-      : super(TransactionOperationType.store, repositoryName, executor,
-            transaction);
+      : super(TransactionOperationType.store, repositoryName, canPropagate,
+            executor, transaction);
 
   @override
   String toString() {
-    return 'TransactionOperation[#$id@#${transaction.id}:store@$repositoryName]{entity: $entity$_commandToString}';
+    return 'TransactionOperation[#$id@#${transaction.id}:store@$repositoryName]{entity: $entity$_commandToString}$durationMsInfo';
   }
 }
 
@@ -4901,14 +5032,14 @@ class TransactionOperationUpdate<O> extends TransactionOperation {
   final O entity;
 
   TransactionOperationUpdate(
-      String repositoryName, Object executor, this.entity,
+      String repositoryName, bool canPropagate, Object executor, this.entity,
       [Transaction? transaction])
-      : super(TransactionOperationType.update, repositoryName, executor,
-            transaction);
+      : super(TransactionOperationType.update, repositoryName, canPropagate,
+            executor, transaction);
 
   @override
   String toString() {
-    return 'TransactionOperation[#$id@#${transaction.id}:update@$repositoryName]{entity: $entity$_commandToString}';
+    return 'TransactionOperation[#$id@#${transaction.id}:update@$repositoryName]{entity: $entity$_commandToString}$durationMsInfo';
   }
 }
 
@@ -4919,12 +5050,12 @@ class TransactionOperationStoreRelationship<O, E> extends TransactionOperation {
   TransactionOperationStoreRelationship(
       String repositoryName, Object executor, this.entity, this.others,
       [Transaction? transaction])
-      : super(TransactionOperationType.storeRelationship, repositoryName,
+      : super(TransactionOperationType.storeRelationship, repositoryName, false,
             executor, transaction);
 
   @override
   String toString() {
-    return 'TransactionOperation[#$id@#${transaction.id}:storeRelationship@$repositoryName]{entity: $entity, other: $others$_commandToString}';
+    return 'TransactionOperation[#$id@#${transaction.id}:storeRelationship@$repositoryName]{entity: $entity, other: $others$_commandToString}$durationMsInfo';
   }
 }
 
@@ -4937,11 +5068,11 @@ class TransactionOperationConstrainRelationship<O, E>
       String repositoryName, Object executor, this.entity, this.others,
       [Transaction? transaction])
       : super(TransactionOperationType.constrainRelationship, repositoryName,
-            executor, transaction);
+            false, executor, transaction);
 
   @override
   String toString() {
-    return 'TransactionOperation[#$id@#${transaction.id}:constrainRelationship@$repositoryName]{entity: $entity, other: $others$_commandToString}';
+    return 'TransactionOperation[#$id@#${transaction.id}:constrainRelationship@$repositoryName]{entity: $entity, other: $others$_commandToString}$durationMsInfo';
   }
 }
 
@@ -4952,26 +5083,27 @@ class TransactionOperationSelectRelationship<O> extends TransactionOperation {
       String repositoryName, Object executor, this.entity,
       [Transaction? transaction])
       : super(TransactionOperationType.selectRelationship, repositoryName,
-            executor, transaction);
+            false, executor, transaction);
 
   @override
   String toString() {
-    return 'TransactionOperation[#$id@#${transaction.id}:selectRelationship@$repositoryName]{entity: $entity$_commandToString}';
+    return 'TransactionOperation[#$id@#${transaction.id}:selectRelationship@$repositoryName]{entity: $entity$_commandToString}$durationMsInfo';
   }
 }
 
 class TransactionOperationSelectRelationships<O> extends TransactionOperation {
+  final String valueRepositoryName;
   final List<O> entities;
 
-  TransactionOperationSelectRelationships(
-      String repositoryName, Object executor, this.entities,
+  TransactionOperationSelectRelationships(String repositoryName,
+      this.valueRepositoryName, Object executor, this.entities,
       [Transaction? transaction])
       : super(TransactionOperationType.selectRelationships, repositoryName,
-            executor, transaction);
+            false, executor, transaction);
 
   @override
   String toString() {
-    return 'TransactionOperation[#$id@#${transaction.id}:selectRelationships@$repositoryName]{entities: $entities$_commandToString}';
+    return 'TransactionOperation[#$id@#${transaction.id}:selectRelationships@$repositoryName->$valueRepositoryName]{entities: $entities$_commandToString}$durationMsInfo';
   }
 }
 
@@ -4979,14 +5111,14 @@ class TransactionOperationDelete<O> extends TransactionOperation {
   final EntityMatcher matcher;
 
   TransactionOperationDelete(
-      String repositoryName, Object executor, this.matcher,
+      String repositoryName, bool canPropagate, Object executor, this.matcher,
       [Transaction? transaction])
-      : super(TransactionOperationType.delete, repositoryName, executor,
-            transaction);
+      : super(TransactionOperationType.delete, repositoryName, canPropagate,
+            executor, transaction);
 
   @override
   String toString() {
-    return 'TransactionOperation[#$id@#${transaction.id}:delete@$repositoryName]{matcher: $matcher$_commandToString}';
+    return 'TransactionOperation[#$id@#${transaction.id}:delete@$repositoryName]{matcher: $matcher$_commandToString}$durationMsInfo';
   }
 }
 
@@ -5052,6 +5184,7 @@ class TransactionEntityProvider
   FutureOr<O?> getEntityByID<O>(id,
       {Type? type, bool sync = false, EntityResolutionRules? resolutionRules}) {
     var oAsync = transaction.getEntityByID<O>(id, type: type, sync: sync);
+
     return oAsync.resolveMapped((o) {
       if (o != null || sync) return o;
 
@@ -5074,7 +5207,8 @@ class TransactionEntityProvider
       if (entityRepository == null) return null;
 
       var sel = entityRepository.selectByID(id,
-          resolutionRules: resolutionRulesResolved);
+          transaction: transaction, resolutionRules: resolutionRulesResolved);
+
       return sel.resolveMapped((o) => o as O?);
     });
   }
@@ -5093,6 +5227,29 @@ abstract class IterableEntityRepository<O extends Object>
   void put(O o);
 
   void remove(O o);
+
+  @override
+  bool hasReferencedEntities() {
+    final fieldsEntity = entityHandler.fieldsWithTypeEntityOrReference();
+    final fieldsEntityRef = entityHandler.fieldsWithEntityReference();
+
+    final fieldsListEntity =
+        entityHandler.fieldsWithTypeListEntityOrReference();
+    final fieldsListEntityRef = entityHandler.fieldsWithEntityReferenceList();
+
+    if (fieldsEntity.isEmpty && fieldsListEntity.isEmpty) return false;
+
+    var fieldsEntityNoRef = fieldsEntity.length - fieldsEntityRef.length;
+    if (fieldsEntityNoRef > 0) return true;
+
+    var fieldsListEntityNoRef =
+        fieldsListEntity.length - fieldsListEntityRef.length;
+    if (fieldsListEntityNoRef > 0) return true;
+
+    if (fieldsEntityRef.isEmpty && fieldsListEntityRef.isNotEmpty) return false;
+
+    return true;
+  }
 
   @override
   FutureOr<int> count(
@@ -5189,7 +5346,10 @@ abstract class IterableEntityRepository<O extends Object>
 
     checkEntityFields(o);
 
-    var op = TransactionOperationStore(name, this, o, transaction);
+    var canPropagate = hasReferencedEntities();
+
+    var op =
+        TransactionOperationStore(name, canPropagate, this, o, transaction);
 
     return ensureReferencesStored(o, transaction: op.transaction)
         .resolveWith(() {